size_t in for

+++ tasks/mpi/korobeinikov_matrix_multiplication_horizontal_scheme_A_vertical_scheme_B/func_tests/main_korobeinikov.cpp
@@ -94,10 +94,10 @@ TEST(matrix_multiplication_horizontal_scheme_A_vertical_scheme_B,
     testMpiTaskSequential.run();
     testMpiTaskSequential.post_processing();
 
-    for (int i = 0; i < right_answer.size(); i++) {
+    for (size_t i = 0; i < right_answer.size(); i++) {
       ASSERT_EQ(right_answer[i], out_seq[i]);
     }
-    for (int i = 0; i < right_answer.size(); i++) {
+    for (size_t i = 0; i < right_answer.size(); i++) {
       ASSERT_EQ(right_answer[i], out_mpi[i]);
     }
     ASSERT_EQ(count_rows_out_seq, count_rows_RA);
@@ -295,10 +295,10 @@ TEST(matrix_multiplication_horizontal_scheme_A_vertical_scheme_B,
     testMpiTaskSequential.run();
     testMpiTaskSequential.post_processing();
 
-    for (int i = 0; i < right_answer.size(); i++) {
+    for (size_t i = 0; i < right_answer.size(); i++) {
       ASSERT_EQ(right_answer[i], out_seq[i]);
     }
-    for (int i = 0; i < right_answer.size(); i++) {
+    for (size_t i = 0; i < right_answer.size(); i++) {
       ASSERT_EQ(right_answer[i], out_mpi[i]);
     }
     ASSERT_EQ(count_rows_out_seq, count_rows_RA);
